+++ src/main/java/cloud/fogbow/fns/core/ApplicationFacade.java
@@ -19,8 +19,8 @@ import cloud.fogbow.fns.constants.SystemConstants;
 import cloud.fogbow.fns.core.drivers.ServiceDriverFactory;
 import cloud.fogbow.fns.core.exceptions.InvalidCidrException;
 import cloud.fogbow.fns.core.model.*;
-import cloud.fogbow.fns.core.serviceconnector.AgentConfiguration;
-import cloud.fogbow.fns.core.serviceconnector.SSAgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.AgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.SSAgentConfiguration;
 import cloud.fogbow.fns.utils.FederatedNetworkUtil;
 import cloud.fogbow.fns.utils.RedirectToRasUtil;
 import cloud.fogbow.ras.api.http.ExceptionResponse;
