+++ src/main/java/cloud/fogbow/fns/core/drivers/ServiceDriver.java
@@ -3,7 +3,7 @@ package cloud.fogbow.fns.core.drivers;
 import cloud.fogbow.common.exceptions.FogbowException;
 import cloud.fogbow.fns.api.parameters.FederatedCompute;
 import cloud.fogbow.fns.core.model.FederatedNetworkOrder;
-import cloud.fogbow.fns.core.serviceconnector.AgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.AgentConfiguration;
 import cloud.fogbow.ras.core.models.UserData;
 
 public interface ServiceDriver {
