+++ src/main/java/cloud/fogbow/fns/core/drivers/vanilla/VanillaServiceDriver.java
@@ -8,8 +8,8 @@ import cloud.fogbow.fns.core.drivers.GeneralServiceDriver;
 import cloud.fogbow.fns.core.model.FederatedNetworkOrder;
 import cloud.fogbow.fns.core.model.MemberConfigurationState;
 import cloud.fogbow.fns.core.model.OrderState;
-import cloud.fogbow.fns.core.serviceconnector.AgentConfiguration;
-import cloud.fogbow.fns.core.serviceconnector.SSAgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.AgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.SSAgentConfiguration;
 import cloud.fogbow.fns.utils.AgentCommunicatorUtil;
 import cloud.fogbow.fns.utils.FederatedComputeUtil;
 import cloud.fogbow.fns.utils.FederatedNetworkUtil;
