+++ src/main/java/cloud/fogbow/fns/core/intercomponent/serviceconnector/AgentConfiguration.java
@@ -1,0 +1,4 @@
+package cloud.fogbow.fns.core.intercomponent.serviceconnector;
+
+public interface AgentConfiguration {
+}
