+++ src/main/java/cloud/fogbow/fns/core/intercomponent/xmpp/requesters/RemoteConfigureAgent.java
@@ -4,8 +4,8 @@ import cloud.fogbow.common.util.GsonHolder;
 import cloud.fogbow.fns.core.intercomponent.xmpp.IqElement;
 import cloud.fogbow.fns.core.intercomponent.xmpp.PacketSenderHolder;
 import cloud.fogbow.fns.core.intercomponent.xmpp.RemoteMethod;
-import cloud.fogbow.fns.core.serviceconnector.AgentConfiguration;
-import cloud.fogbow.fns.core.serviceconnector.SSAgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.AgentConfiguration;
+import cloud.fogbow.fns.core.intercomponent.serviceconnector.SSAgentConfiguration;
 import cloud.fogbow.ras.core.intercomponent.xmpp.XmppErrorConditionToExceptionTranslator;
 import org.dom4j.Element;
 import org.xmpp.packet.IQ;
