--- src/main/java/cloud/fogbow/fns/core/serviceconnector/AgentConfiguration.java
@@ -1,4 +1,0 @@
-package cloud.fogbow.fns.core.serviceconnector;
-
-public interface AgentConfiguration {
-}
